+++ example/test/repositories/people_repository_test.dart
@@ -1,7 +1,6 @@
 import 'package:clock/clock.dart';
 import 'package:cornerstone/cornerstone.dart';
 import 'package:dartz/dartz.dart';
-import 'package:equatable/equatable.dart';
 import 'package:example/data_sources/people_data_source.dart';
 import 'package:example/entities/person.dart';
 import 'package:example/repositories/people_repository.dart';
@@ -186,7 +185,6 @@ void main() {
       });
 
       test('should set loaded data to repo and return the data', () async {
-        EquatableConfig.stringify = true;
         when(box.toMap()).thenReturn({...snapJsonFixture, 'isSaved': true});
         final result = await repo.load();
         expect(