+++ lib/cornerstone.dart
@@ -1,6 +1,4 @@
 library cornerstone;
 
-export 'src/data_source.dart';
-export 'src/failure.dart';
-export 'src/repository.dart';
-export 'src/use_case.dart';
+export './src/core/barrel_export_core.dart';
+export './src/utilities/barrel_export_utilities.dart';